--- overlaps2_test.go
@@ -1,532 +1,0 @@
-// Copyright (c) 2025 Karl Gaissmaier
-// SPDX-License-Identifier: MIT
-
-package bart
-
-import (
-	"math/rand/v2"
-	"net/netip"
-	"testing"
-)
-
-// expectedOverlaps computes a slow O(n*m) ground-truth overlap between two sets.
-func expectedOverlaps(as, bs []netip.Prefix) bool {
-	for _, a := range as {
-		for _, b := range bs {
-			if a.Overlaps(b) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// expectedOverlapsPrefix computes whether any prefix in set as overlaps q.
-func expectedOverlapsPrefix(as []netip.Prefix, q netip.Prefix) bool {
-	for _, a := range as {
-		if a.Overlaps(q) {
-			return true
-		}
-	}
-	return false
-}
-
-// FuzzOverlaps validates Overlaps(o) vs ground-truth for all table types.
-func FuzzOverlaps(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(12345), 50, 50)
-	f.Add(uint64(67890), 150, 75)
-	f.Add(uint64(22222), 300, 300)
-
-	f.Fuzz(func(t *testing.T, seed uint64, n1, n2 int) {
-		if n1 < 1 || n1 > 1000 || n2 < 1 || n2 > 1000 {
-			t.Skip("counts out of range")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 42))
-		aItems := randomPrefixes(prng, n1) // invariant: no duplicates
-		bItems := randomPrefixes(prng, n2)
-
-		aPfx := make([]netip.Prefix, len(aItems))
-		for i, it := range aItems {
-			aPfx[i] = it.pfx
-		}
-		bPfx := make([]netip.Prefix, len(bItems))
-		for i, it := range bItems {
-			bPfx[i] = it.pfx
-		}
-
-		gt := expectedOverlaps(aPfx, bPfx)
-
-		// Table
-		{
-			t1 := new(Table[int])
-			t2 := new(Table[int])
-			for _, it := range aItems {
-				t1.Insert(it.pfx, it.val)
-			}
-			for _, it := range bItems {
-				t2.Insert(it.pfx, it.val)
-			}
-			if got := t1.Overlaps(t2); got != gt {
-				t.Fatalf("Table.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// Fast
-		{
-			f1 := new(Fast[int])
-			f2 := new(Fast[int])
-			for _, it := range aItems {
-				f1.Insert(it.pfx, it.val)
-			}
-			for _, it := range bItems {
-				f2.Insert(it.pfx, it.val)
-			}
-			if got := f1.Overlaps(f2); got != gt {
-				t.Fatalf("Fast.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// liteTable (payload-less semantics—overlap depends only on prefixes)
-		{
-			lt1 := new(liteTable[int])
-			lt2 := new(liteTable[int])
-			for _, it := range aItems {
-				lt1.Insert(it.pfx, 0)
-			}
-			for _, it := range bItems {
-				lt2.Insert(it.pfx, 0)
-			}
-			if got := lt1.Overlaps(lt2); got != gt {
-				t.Fatalf("liteTable.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// Lite (prefix-only)
-		{
-			L1 := new(Lite)
-			L2 := new(Lite)
-			for _, it := range aItems {
-				L1.Insert(it.pfx)
-			}
-			for _, it := range bItems {
-				L2.Insert(it.pfx)
-			}
-			if got := L1.Overlaps(L2); got != gt {
-				t.Fatalf("Lite.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-	})
-}
-
-// FuzzOverlapsPrefix validates OverlapsPrefix(q) vs ground-truth for all table types.
-// For each query, it checks that OverlapsPrefix matches any(a.Overlaps(q)).
-func FuzzOverlapsPrefix(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(111), 200, 60)
-	f.Add(uint64(222), 600, 120)
-	f.Add(uint64(333), 1200, 180)
-
-	f.Fuzz(func(t *testing.T, seed uint64, n, nq int) {
-		if n < 1 || n > 4000 || nq < 1 || nq > 400 {
-			t.Skip("counts out of range")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 7))
-		items := randomPrefixes(prng, n)  // table contents
-		query := randomPrefixes(prng, nq) // queries, use .pfx
-
-		contents := make([]netip.Prefix, len(items))
-		for i, it := range items {
-			contents[i] = it.pfx
-		}
-
-		// Table
-		{
-			tbl := new(Table[int])
-			for _, it := range items {
-				tbl.Insert(it.pfx, it.val)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := tbl.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Table.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// Fast
-		{
-			fst := new(Fast[int])
-			for _, it := range items {
-				fst.Insert(it.pfx, it.val)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := fst.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Fast.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// liteTable (presence only)
-		{
-			lt := new(liteTable[int])
-			for _, it := range items {
-				lt.Insert(it.pfx, 0)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := lt.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("liteTable.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// Lite (prefix-only)
-		{
-			L := new(Lite)
-			for _, it := range items {
-				L.Insert(it.pfx)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := L.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Lite.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-	})
-}
-
-// TestOverlapsDeterministic_EdgeCases exercises thorough edge coverage for Overlaps across all table types:
-// - IPv6: siblings (/33 under /32), default route (::/0), host routes (/128), /127 semantics
-// - IPv4: default route (0.0.0.0/0), host routes (/32), /31 semantics, adjacent siblings
-// - Mixed-family disjointness
-func TestOverlapsDeterministic_EdgeCases(t *testing.T) {
-	t.Parallel()
-
-	type pair struct {
-		name  string
-		pfxsA []string
-		pfxsB []string
-		want  bool // expected Overlaps(A,B)
-	}
-
-	cases := []pair{
-		// IPv6: two halves of a /32 are siblings and do not overlap
-		{
-			name:  "IPv6_siblings_/33",
-			pfxsA: []string{"2001:db8::/33"},
-			pfxsB: []string{"2001:db8:8000::/33"},
-			want:  false,
-		},
-		// IPv6: default route overlaps everything in IPv6 space
-		{
-			name:  "IPv6_default_overlaps_db8_32",
-			pfxsA: []string{"::/0"},
-			pfxsB: []string{"2001:db8::/32"},
-			want:  true,
-		},
-		// IPv6: host vs default
-		{
-			name:  "IPv6_host_vs_default_zero",
-			pfxsA: []string{"::/128"},
-			pfxsB: []string{"::/0"},
-			want:  true,
-		},
-		// IPv6: max host vs default
-		{
-			name:  "IPv6_max_host_vs_default",
-			pfxsA: []string{"ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128"},
-			pfxsB: []string{"::/0"},
-			want:  true,
-		},
-		// IPv6: exact host
-		{
-			name:  "IPv6_exact_host",
-			pfxsA: []string{"2001:db8::1/128"},
-			pfxsB: []string{"2001:db8::1/128"},
-			want:  true,
-		},
-		// IPv6: adjacent /64 siblings
-		{
-			name:  "IPv6_adjacent_/64",
-			pfxsA: []string{"2001:db8::/64"},
-			pfxsB: []string{"2001:db8:0:1::/64"},
-			want:  false,
-		},
-		// IPv6: /127 covers exactly two hosts
-		{
-			name:  "IPv6_/127_covers_two_hosts",
-			pfxsA: []string{"2001:db8::/127"},
-			pfxsB: []string{"2001:db8::/128", "2001:db8::1/128"},
-			want:  true,
-		},
-		// IPv6: /127 excludes next host outside its range
-		{
-			name:  "IPv6_/127_excludes_next_host",
-			pfxsA: []string{"2001:db8::/127"},
-			pfxsB: []string{"2001:db8::2/128"},
-			want:  false,
-		},
-		// IPv6: adjacent /127 ranges do not overlap
-		{
-			name:  "IPv6_adjacent_/127",
-			pfxsA: []string{"2001:db8::/127"},
-			pfxsB: []string{"2001:db8::2/127"},
-			want:  false,
-		},
-		// IPv6: /127 overlapped by its /126 supernet
-		{
-			name:  "IPv6_/127_vs_/126",
-			pfxsA: []string{"2001:db8::/127"},
-			pfxsB: []string{"2001:db8::/126"},
-			want:  true,
-		},
-		// IPv4: default route overlaps everything in IPv4 space
-		{
-			name:  "IPv4_default_vs_host",
-			pfxsA: []string{"0.0.0.0/0"},
-			pfxsB: []string{"255.255.255.255/32"},
-			want:  true,
-		},
-		// IPv4: adjacent /25 siblings do not overlap
-		{
-			name:  "IPv4_adjacent_/25",
-			pfxsA: []string{"10.0.0.0/25"},
-			pfxsB: []string{"10.0.0.128/25"},
-			want:  false,
-		},
-		// IPv4: exact host
-		{
-			name:  "IPv4_exact_host",
-			pfxsA: []string{"192.168.1.1/32"},
-			pfxsB: []string{"192.168.1.1/32"},
-			want:  true,
-		},
-		// IPv4: /31 covers exactly two hosts
-		{
-			name:  "IPv4_/31_contains_two_hosts",
-			pfxsA: []string{"192.0.2.0/31"},
-			pfxsB: []string{"192.0.2.0/32", "192.0.2.1/32"},
-			want:  true,
-		},
-		// IPv4: /31 excludes next host outside its range
-		{
-			name:  "IPv4_/31_excludes_next_host",
-			pfxsA: []string{"192.0.2.0/31"},
-			pfxsB: []string{"192.0.2.2/32"},
-			want:  false,
-		},
-		// IPv4: adjacent /31 ranges do not overlap
-		{
-			name:  "IPv4_adjacent_/31",
-			pfxsA: []string{"192.0.2.0/31"},
-			pfxsB: []string{"192.0.2.2/31"},
-			want:  false,
-		},
-		// IPv4: /31 overlapped by its /30 supernet
-		{
-			name:  "IPv4_/31_vs_/30",
-			pfxsA: []string{"192.0.2.0/31"},
-			pfxsB: []string{"192.0.2.0/30"},
-			want:  true,
-		},
-		// Mixed-family: IPv4 and IPv6 do not overlap
-		{
-			name:  "Mixed_v4_vs_v6",
-			pfxsA: []string{"10.0.0.0/8"},
-			pfxsB: []string{"2001:db8::/32"},
-			want:  false,
-		},
-		// Multi-entry: at least one overlapping pair exists
-		{
-			name:  "Multi_entry_overlap",
-			pfxsA: []string{"10.0.0.0/8", "172.16.0.0/12"},
-			pfxsB: []string{"192.168.0.0/16", "10.64.0.0/10"},
-			want:  true,
-		},
-		// Multi-entry: completely disjoint sets
-		{
-			name:  "Multi_entry_disjoint",
-			pfxsA: []string{"10.0.0.0/16", "172.16.0.0/12"},
-			pfxsB: []string{"10.1.0.0/16", "203.0.113.0/24"},
-			want:  false,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run("Table_"+tc.name, func(t *testing.T) {
-			t.Parallel()
-			a := new(Table[int])
-			b := new(Table[int])
-			for i, s := range tc.pfxsA {
-				a.Insert(netip.MustParsePrefix(s), i)
-			}
-			for i, s := range tc.pfxsB {
-				b.Insert(netip.MustParsePrefix(s), i)
-			}
-			if got := a.Overlaps(b); got != tc.want {
-				t.Fatalf("Table.Overlaps: want %v, got %v (A=%v B=%v)", tc.want, got, tc.pfxsA, tc.pfxsB)
-			}
-		})
-
-		t.Run("Fast_"+tc.name, func(t *testing.T) {
-			t.Parallel()
-			a := new(Fast[int])
-			b := new(Fast[int])
-			for i, s := range tc.pfxsA {
-				a.Insert(netip.MustParsePrefix(s), i)
-			}
-			for i, s := range tc.pfxsB {
-				b.Insert(netip.MustParsePrefix(s), i)
-			}
-			if got := a.Overlaps(b); got != tc.want {
-				t.Fatalf("Fast.Overlaps: want %v, got %v (A=%v B=%v)", tc.want, got, tc.pfxsA, tc.pfxsB)
-			}
-		})
-
-		t.Run("liteTable_"+tc.name, func(t *testing.T) {
-			t.Parallel()
-			a := new(liteTable[int])
-			b := new(liteTable[int])
-			for i, s := range tc.pfxsA {
-				a.Insert(netip.MustParsePrefix(s), i)
-			}
-			for i, s := range tc.pfxsB {
-				b.Insert(netip.MustParsePrefix(s), i)
-			}
-			if got := a.Overlaps(b); got != tc.want {
-				t.Fatalf("liteTable.Overlaps: want %v, got %v (A=%v B=%v)", tc.want, got, tc.pfxsA, tc.pfxsB)
-			}
-		})
-
-		t.Run("Lite_"+tc.name, func(t *testing.T) {
-			t.Parallel()
-			a := new(Lite)
-			b := new(Lite)
-			for _, s := range tc.pfxsA {
-				a.Insert(netip.MustParsePrefix(s))
-			}
-			for _, s := range tc.pfxsB {
-				b.Insert(netip.MustParsePrefix(s))
-			}
-			if got := a.Overlaps(b); got != tc.want {
-				t.Fatalf("Lite.Overlaps: want %v, got %v (A=%v B=%v)", tc.want, got, tc.pfxsA, tc.pfxsB)
-			}
-		})
-	}
-}
-
-// TestOverlapsPrefixDeterministic_EdgeCases exercises corner queries against a mixed content set
-// for all table types. It includes: default routes, host routes, IPv6 /127, IPv4 /31, mixed-family.
-func TestOverlapsPrefixDeterministic_EdgeCases(t *testing.T) {
-	t.Parallel()
-
-	// Mixed IPv4/IPv6 contents, covering host routes and small subnets.
-	contents := []string{
-		"10.0.0.0/8",
-		"192.0.2.0/31",
-		"203.0.113.128/25",
-		"2001:db8::/32",
-		"2001:db8::/127",
-		"::/128",
-	}
-
-	// Queries and expected OverlapsPrefix results with the content above.
-	queries := []struct {
-		probes string
-		want   bool
-	}{
-		// IPv4 /31 semantics and supernets
-		{"192.0.2.0/32", true},
-		{"192.0.2.1/32", true},
-		{"192.0.2.2/32", false},
-		{"192.0.2.0/31", true},
-		{"192.0.2.0/30", true},
-		// Another IPv4 network
-		{"203.0.113.255/32", true}, // within 203.0.113.128/25
-		{"203.0.113.0/25", false},  // different /25
-		{"0.0.0.0/0", true},        // supernet of all IPv4
-
-		// IPv6 /127 and host routes
-		{"2001:db8::/127", true},
-		{"2001:db8::/128", true},
-		{"2001:db8::1/128", true},
-		{"2001:db8::2/128", true},
-		{"2001:db9::1/128", false},
-		{"::/0", true},     // supernet of all IPv6
-		{"::1/128", false}, // different host than ::/128
-	}
-
-	var cPfx []netip.Prefix
-	for _, s := range contents {
-		cPfx = append(cPfx, netip.MustParsePrefix(s))
-	}
-
-	// Table
-	t.Run("Table", func(t *testing.T) {
-		t.Parallel()
-		tbl := new(Table[int])
-		for i, p := range cPfx {
-			tbl.Insert(p, i)
-		}
-		for _, q := range queries {
-			got := tbl.OverlapsPrefix(netip.MustParsePrefix(q.probes))
-			if got != q.want {
-				t.Fatalf("Table.OverlapsPrefix(%s): want %v, got %v", q.probes, q.want, got)
-			}
-		}
-	})
-
-	// Fast
-	t.Run("Fast", func(t *testing.T) {
-		t.Parallel()
-		ft := new(Fast[int])
-		for i, p := range cPfx {
-			ft.Insert(p, i)
-		}
-		for _, q := range queries {
-			got := ft.OverlapsPrefix(netip.MustParsePrefix(q.probes))
-			if got != q.want {
-				t.Fatalf("Fast.OverlapsPrefix(%s): want %v, got %v", q.probes, q.want, got)
-			}
-		}
-	})
-
-	// liteTable
-	t.Run("liteTable", func(t *testing.T) {
-		t.Parallel()
-		lt := new(liteTable[int])
-		for i, p := range cPfx {
-			lt.Insert(p, i)
-		}
-		for _, q := range queries {
-			got := lt.OverlapsPrefix(netip.MustParsePrefix(q.probes))
-			if got != q.want {
-				t.Fatalf("liteTable.OverlapsPrefix(%s): want %v, got %v", q.probes, q.want, got)
-			}
-		}
-	})
-
-	// Lite
-	t.Run("Lite", func(t *testing.T) {
-		t.Parallel()
-		L := new(Lite)
-		for _, p := range cPfx {
-			L.Insert(p)
-		}
-		for _, q := range queries {
-			got := L.OverlapsPrefix(netip.MustParsePrefix(q.probes))
-			if got != q.want {
-				t.Fatalf("Lite.OverlapsPrefix(%s): want %v, got %v", q.probes, q.want, got)
-			}
-		}
-	})
-}
